backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.4
sqlalchemy==2.0.19
asyncpg==0.29.0
psycopg2-binary==2.9.6
alembic==1.10.4
pydantic==2.8.2
python-dotenv==1.0.0
pytest==8.3.2
email-validator==2.2.0
pytest-asyncio==0.23.8